INS-3088: new simple GetIndex procedure instead of EnsureIndex.Result
