+++ ledger/light/handle/get_object.go
@@ -69,17 +69,25 @@ func (s *GetObject) Present(ctx context.Context, f flow.Flow) error {
 		return err
 	}
 
-	idx := proc.NewEnsureIndexWM(msg.ObjectID, objJetID, s.meta)
-	s.dep.EnsureIndex(idx)
-	if err := f.Procedure(ctx, idx, false); err != nil {
+	ensureIdx := proc.NewEnsureIndexWM(msg.ObjectID, objJetID, s.meta)
+	s.dep.EnsureIndex(ensureIdx)
+	if err := f.Procedure(ctx, ensureIdx, false); err != nil {
 		return err
 	}
 
-	if idx.Result.Lifeline.StateID == record.StateDeactivation {
+	getIdx := proc.NewGetIndex(msg.ObjectID)
+	s.dep.GetIndexWM(getIdx)
+	if err := f.Procedure(ctx, getIdx, false); err != nil {
+		return err
+	}
+
+	idx := getIdx.Result.Lifeline
+
+	if idx.StateID == record.StateDeactivation {
 		return errors.New("object is deactivated")
 	}
 
-	send := proc.NewSendObject(s.meta, msg.ObjectID, idx.Result.Lifeline)
+	send := proc.NewSendObject(s.meta, msg.ObjectID, idx)
 	s.dep.SendObject(send)
 	return f.Procedure(ctx, send, false)
 }
