+++ ledger/light/proc/proc.go
@@ -22,6 +22,7 @@ type Dependencies struct {
 	WaitHot          func(*WaitHot)
 	WaitHotWM        func(*WaitHotWM)
 	GetIndex         func(*EnsureIndex)
+	GetIndexWM       func(*GetIndex)
 	EnsureIndex      func(*EnsureIndexWM)
 	SendObject       func(*SendObject)
 	GetCode          func(*GetCode)
@@ -52,6 +53,7 @@ func NewDependenciesMock() *Dependencies {
 		CheckJet:         func(*CheckJet) {},
 		WaitHot:          func(*WaitHot) {},
 		WaitHotWM:        func(*WaitHotWM) {},
+		GetIndexWM:       func(*GetIndex) {},
 		GetIndex:         func(*EnsureIndex) {},
 		EnsureIndex:      func(*EnsureIndexWM) {},
 		SendObject:       func(*SendObject) {},
